feat: added lock and unlock to the door
Once the key has been clicked on, the Unlock will be called creating the locked boolean to go to false. In doing so, when the player clicks on the Door it will open up causing the door to rotate up allowing the player into the temple. If the key has not been clicked on, then the door will play the locked door sound and not open.